Add files via upload
Takes a Drawing number list with a collapsible top-down cell structure in an excel sheet and creates a matching Folder structure in a user given directory. Limitation is that this code is specific to the excel sheet being used. Could only be generalized if the drawing number list format is consistent which usually it is subjective to the drafter.