Add files via upload
Utilized Python to automate a process to gather information about an antiquated shared drive that includes number of folders and sub-folders, number of files, and total folder size at each folder/project level. The result was an excel spreadsheet with 29,000 datapoints that will facilitate the migration of that legacy data.